DataStore module added to persist required data. Location monitor added using fused location. Permissions request implemented.

+++ core/network/src/main/java/com/arch/network/webhook/model/UserLocationBody.kt
@@ -4,8 +4,8 @@ import com.google.gson.annotations.SerializedName
 
 data class UserLocationBody(
     @SerializedName("uid") val userId: String,
-    @SerializedName("lat") val latitude: Float,
-    @SerializedName("lng") val longitude: Float,
+    @SerializedName("lat") val latitude: Double,
+    @SerializedName("lng") val longitude: Double,
     @SerializedName("accuracy") val accuracy: Float,
     @SerializedName("timestamp") val timestamp: Double
 )
